pyproject.toml
[tool.poetry]
name = "nextx.repository"
version = "0.1.0"
description = "A repository interface and a companion stub implementation of NoSQL MongoDb"
authors = ["adriangs1996 <adriangonzalezsanchez1996@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
beanie = "^1.10.4"
fastapi = "^0.75.1"
python-jose = "^3.3.0"
aioredis = "^2.0.1"
python-decouple = "^3.6"

[tool.poetry.dev-dependencies]
black = "^22.3.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
